src/requirements.txt
black>=24.4.2
isort>=5.13.2
dj-database-url>=2.2.0
Django>=5.0.7
cryptography>=43.0.0
django-allauth>=0.63.6
dj-rest-auth>=6.0.0
django-cors-headers>=4.4.0
drf-yasg>=1.21.7
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
pillow>=10.4.0
django-cleanup>=8.1.0
django-cloudinary-storage>=0.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
gunicorn>=22.0.0
whitenoise>=6.7.0
setuptools>=71.1.0
django-extensions>=3.2.3
pydotplus>=2.0.2
